tech:move checkCampaignBelongsToCombinedCourse from shared context to Prescription campaigns context
