+++ api/tests/shared/acceptance/application/security-pre-handlers_test.js
@@ -1,5 +1,4 @@
 import { createServer } from '../../../../server.js';
-import { CombinedCourseBlueprint } from '../../../../src/quest/domain/models/combined-course-blueprints/entities/CombinedCourseBlueprint.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
 import { ORGANIZATION_FEATURE } from '../../../../src/shared/constants.js';
 import { Membership } from '../../../../src/shared/domain/models/Membership.js';
@@ -22,53 +21,6 @@ describe('Acceptance | Application | SecurityPreHandlers', function () {
 
   beforeEach(async function () {
     server = await createServer();
-  });
-
-  describe('#checkCampaignBelongsToCombinedCourse', function () {
-    let campaignId;
-    beforeEach(async function () {
-      server.route({
-        method: 'GET',
-        path: '/api/pate-de-campagne/{campaignId}',
-        handler: (r, h) => h.response({}).code(200),
-        config: {
-          auth: false,
-          pre: [
-            {
-              method: securityPreHandlers.checkCampaignBelongsToCombinedCourse,
-            },
-          ],
-        },
-      });
-
-      const organizationId = databaseBuilder.factory.buildOrganization().id;
-      campaignId = databaseBuilder.factory.buildCampaign({ organizationId }).id;
-      const { id: questId } = databaseBuilder.factory.buildQuestForCombinedCourse({
-        successRequirements: [CombinedCourseBlueprint.buildRequirementForCombinedCourse({ campaignId }).toDTO()],
-      });
-      databaseBuilder.factory.buildCombinedCourse({
-        code: 'ABCDE1234',
-        name: 'Mon parcours Combiné',
-        organizationId,
-        questId,
-      });
-      await databaseBuilder.commit();
-    });
-
-    it('should return a well formed JSON API error when campaign belongs to a combined course', async function () {
-      // given
-      const options = {
-        method: 'GET',
-        url: `/api/pate-de-campagne/${campaignId}`,
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      expect(response.result.errors[0].code).equal('CAMPAIGN_BELONGS_TO_COMBINED_COURSE');
-    });
   });
 
   describe('#checkAdminMemberHasRoleSuperAdmin', function () {
